orchestrator/requirements.txt
# Runtime dependencies — installed in the Docker image.
# Dev/test tools live in requirements-dev.txt and are NOT installed in Docker.
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
anthropic>=0.80.0
openai>=2.0.0
httpx>=0.27.0
# Web Push (Phase 4) — outbound browser notifications via subscription rows + VAPID.
pywebpush>=2.0.0,<3
pyyaml>=6.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
# packaging is used by services/capability_registry.py for min_core_version
# comparison via packaging.version.Version. Universally shipped via pip,
# declared explicitly so we don't rely on transitive installation.
packaging>=21.0
qdrant-client>=1.12.0
psycopg2-binary>=2.9.0
pdfminer.six>=20231228
tiktoken>=0.10.0
sentence-transformers>=5.0.0
pytesseract>=0.3.10
python-docx>=1.0.0
pdf2image>=1.17.0
watchdog>=5.0.0
psutil>=6.0.0
pyjwt>=2.8.0
# structlog for the structured-logging foundation (chunk:
# structured_audit_logging — see docs/structured-logging.md). MIT OR
# Apache-2.0 dual licence; pure-python; integrates with stdlib so
# existing logging.getLogger(__name__) call sites continue to work.
structlog>=25.0
# argon2-cffi for password hashing in services/users.py (family-LAN auth, ADR family_lan_multi_user).
# Apache-2.0; PasswordHasher().verify() raises VerifyMismatchError / InvalidHashError —
# both subclasses of VerificationError which is what services/users.py catches.
argon2-cffi>=23.1.0,<24
# email-validator is a runtime dependency of pydantic v2's EmailStr; required by
# orchestrator/models/auth.py. BSD-licensed; pure-python.
email-validator>=2.0.0,<3
feedparser>=6.0.0
trafilatura>=2.0.0
apscheduler>=3.10.0
caldav>=1.3.0
falkordb>=1.0.0
# mcp 1.10.0 was the first stable streamable-HTTP release; the orchestrator
# uses it via mcp_server.py to expose Lumogis as an MCP server at /mcp.
# Pulls pyjwt[crypto] (and therefore the `cryptography` wheel ~5MB).
mcp>=1.10.0
# splink major 4 — pinned 2026-04-15
splink>=4.0,<5
# duckdb major 1 — pinned 2026-04-15 (Splink DuckDB backend, in-process, no separate service)
duckdb>=1.0,<2
